feat(server): make the rate limit the deployment's, not each replica's
Two replicas behind a load balancer sharing a database is the first
architecture anyone deploying this reaches for, and nothing in the repository
had ever run it. So the first commit here is the measurement, not the fix:
`tests/multi_replica.rs` runs two handlers over one PostgreSQL instance and
asserts one claim per test.

Three of the four answers were already right, which is worth saying as
plainly as the one that was not:

  * a task created on one replica is readable on another — the store holds it
  * a subscription on the *other* replica still terminates when the task
    finishes, reporting the terminal state, because the reattach hook polls
    the shared store. That is §3.1.6 / STREAM-SUB-002 satisfied across
    replicas, which had been assumed rather than checked.
  * intermediate events do not cross replicas — event queues are per-process.
    Architectural rather than broken, and now pinned by a test so that a
    future pub/sub queue has to rewrite the claim rather than silently
    contradict the docs.

The fourth is the defect. `RateLimitInterceptor` counts in a process-local
map, so each replica admits the full configured rate and N replicas admit N
times it: two limiters at 5 per window admit 10, measured. For a limiter
protecting an upstream with a real quota that is the whole point defeated,
and it fails weakest in exactly the deployment that needs it most.

`RateLimitCounter` is the fix — one method, "count this and tell me the
total" — with `PostgresRateLimitCounter` shipped because an agent needing a
shared limiter is almost certainly already sharing a store, and a second
piece of infrastructure to fix a counter is a poor trade. Anyone already
running Redis implements the trait against it in a few lines, which is why
this is a trait rather than a dependency.

Three decisions worth stating rather than leaving to be discovered:

  * **Atomicity is the contract.** The count is one
    `INSERT .. ON CONFLICT DO UPDATE .. RETURNING`, so two replicas hitting
    the same row get 1 and 2. A read-then-write loses increments under
    precisely the concurrency this exists for — the one bug that would make a
    shared counter worse than none, because it would look like it worked.
  * **An unreachable counter degrades to local counting**, so the failure
    mode is exactly the per-process behaviour the deployment already had.
    Failing closed would make adopting a shared limiter a reliability
    regression; failing open would remove the limit at the moment an attacker
    with database access most wants it gone.
  * **The counter's own pool runs `synchronous_commit = off`**, which is
    598us -> 232us per request. A rate-limit count describes one window and is
    swept shortly after, so a crash can at worst forget part of a caller's
    spent budget. `from_pool` deliberately does *not* apply it: that pool is
    usually the task store's, and quietly making a task store non-durable to
    speed up a counter would be an appalling thing to do behind a caller's
    back.

The cost is real and is in the docs rather than buried: 0.2us in-process
against 232us through PostgreSQL on loopback, where a whole JSON-RPC request
measures ~195us. A shared counter roughly doubles a request. That is why it
is opt-in.

Four injections, and two found missing tests rather than confirming present
ones:

  * making the counter ignore the caller key — caught, once the injection
    actually applied; the first attempt did not match and reported a pass.
  * a non-atomic upsert — caught.
  * ignoring the limit on the shared path — caught.
  * removing the fallback — **not** caught. The test built a "down" counter by
    dropping its database, and the drop silently failed because pools were
    open, so the counter kept working and the test passed while exercising
    nothing. Postgres needs `WITH (FORCE)` here, which the existing store
    suite already knew and this file did not; seven scratch databases were
    leaking as a result. The fallback is now tested against a stub counter
    that always fails, where it belongs — it is interceptor behaviour, not
    PostgreSQL behaviour — and it runs in ordinary CI rather than only in the
    database job.

Also: a book chapter, because the gap included having no guidance at all.
Its examples compile rather than being `ignore`d, which needed `postgres`
adding to book-tests — a crate whose own comment says "all features".

rate_limit.rs comes *off* the file-length baseline in the process: 1,272
lines down to 496, after the counter, caller identity and the tests moved to
their own files.
