+++ .file-length-baseline
@@ -32,6 +32,20 @@
 #   listing, pagination, filters and concurrency, each against real SQLite —
 #   and splitting it by feature would scatter the fixtures they share.
 #
+# rate_limit.rs came OFF this list when the shared counter was added, which is
+# the direction it is supposed to move. It was 1,272 lines; the implementation
+# is now 496 in rate_limit/mod.rs, under the limit, after the counter went to
+# shared.rs, caller identity to identity.rs, and the tests to their own files.
+# rate_limit/tests.rs (833) is the extracted suite, recorded on the same
+# reasoning as the two above.
+#
+# tests/multi_replica.rs (531) is a new file and the one entry here that is not
+# a split. It is 31 lines over, and most of its length is the paragraph above
+# each test saying which multi-replica claim that test pins and why anyone
+# should care — for a file whose entire purpose is to replace assumptions about
+# two-replica behaviour with measurements, deleting that prose to save 31 lines
+# would remove the part that makes it useful.
+#
 # Regenerating with --update drops every comment above. If you run it, put them
 # back.
 benches/benches/advanced_scenarios.rs
@@ -70,7 +84,7 @@ crates/a2a-protocol-server/src/handler/messaging/tests.rs
 crates/a2a-protocol-server/src/handler/push_config.rs
 crates/a2a-protocol-server/src/push/config_store.rs
 crates/a2a-protocol-server/src/push/sender.rs
-crates/a2a-protocol-server/src/rate_limit.rs
+crates/a2a-protocol-server/src/rate_limit/tests.rs
 crates/a2a-protocol-server/src/store/sqlite_store/mod.rs
 crates/a2a-protocol-server/src/store/sqlite_store/tests.rs
 crates/a2a-protocol-server/src/store/task_store/in_memory/mod.rs
@@ -85,6 +99,7 @@ crates/a2a-protocol-server/tests/dispatch_tests/jsonrpc.rs
 crates/a2a-protocol-server/tests/dispatch_tests/rest.rs
 crates/a2a-protocol-server/tests/event_queue_tests.rs
 crates/a2a-protocol-server/tests/jsonrpc_dispatch_coverage_tests.rs
+crates/a2a-protocol-server/tests/multi_replica.rs
 crates/a2a-protocol-server/tests/postgres_store_tests.rs
 crates/a2a-protocol-server/tests/push_sender_tests.rs
 crates/a2a-protocol-server/tests/tenant_sqlite_store_tests.rs
