+++ book/src/SUMMARY.md
@@ -55,6 +55,7 @@
     - [Bugs Found & Fixed](./deployment/dogfooding-bugs.md)
     - [Test Coverage Matrix](./deployment/dogfooding-tests.md)
 - [Production Hardening](./deployment/production.md)
+- [Running More Than One Replica](./deployment/horizontal-scaling.md)
 - [GitHub Pages & CI/CD](./deployment/cicd.md)
 
 # Reference
